+++ crates/miden-testing/src/kernel_tests/tx/test_account_delta.rs
@@ -3,10 +3,8 @@ use std::collections::BTreeMap;
 use std::string::String;
 
 use anyhow::Context;
-use miden_lib::testing::account_component::MockAccountComponent;
-use miden_lib::utils::CodeBuilder;
-use miden_objects::account::delta::AccountUpdateDetails;
-use miden_objects::account::{
+use miden_protocol::account::delta::AccountUpdateDetails;
+use miden_protocol::account::{
     Account,
     AccountBuilder,
     AccountDelta,
@@ -19,9 +17,9 @@ use miden_objects::account::{
     StorageSlotDelta,
     StorageSlotName,
 };
-use miden_objects::asset::{Asset, AssetVault, FungibleAsset, NonFungibleAsset};
-use miden_objects::note::{Note, NoteExecutionHint, NoteTag, NoteType};
-use miden_objects::testing::account_id::{
+use miden_protocol::asset::{Asset, AssetVault, FungibleAsset, NonFungibleAsset};
+use miden_protocol::note::{Note, NoteExecutionHint, NoteTag, NoteType};
+use miden_protocol::testing::account_id::{
     ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_1,
     ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_2,
     ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_3,
@@ -29,17 +27,19 @@ use miden_objects::testing::account_id::{
     ACCOUNT_ID_SENDER,
     AccountIdBuilder,
 };
-use miden_objects::testing::asset::NonFungibleAssetBuilder;
-use miden_objects::testing::constants::{
+use miden_protocol::testing::asset::NonFungibleAssetBuilder;
+use miden_protocol::testing::constants::{
     CONSUMED_ASSET_1_AMOUNT,
     CONSUMED_ASSET_3_AMOUNT,
     FUNGIBLE_ASSET_AMOUNT,
     NON_FUNGIBLE_ASSET_DATA,
     NON_FUNGIBLE_ASSET_DATA_2,
 };
-use miden_objects::testing::storage::{MOCK_MAP_SLOT, MOCK_VALUE_SLOT0};
-use miden_objects::transaction::TransactionScript;
-use miden_objects::{EMPTY_WORD, Felt, FieldElement, LexicographicWord, Word, ZERO};
+use miden_protocol::testing::storage::{MOCK_MAP_SLOT, MOCK_VALUE_SLOT0};
+use miden_protocol::transaction::TransactionScript;
+use miden_protocol::{EMPTY_WORD, Felt, FieldElement, LexicographicWord, Word, ZERO};
+use miden_standards::code_builder::CodeBuilder;
+use miden_standards::testing::account_component::MockAccountComponent;
 use miden_tx::LocalTransactionProver;
 use rand::{Rng, SeedableRng};
 use rand_chacha::ChaCha20Rng;
@@ -643,7 +643,7 @@ async fn asset_and_storage_delta() -> anyhow::Result<()> {
 
             # move an asset to the created note to partially deplete fungible asset balance
             swapw dropw push.{REMOVED_ASSET}
-            call.::miden::contracts::wallets::basic::move_asset_to_note
+            call.::miden::standards::wallets::basic::move_asset_to_note
             # => [ASSET, note_idx, pad(11)]
 
             # clear the stack
@@ -659,8 +659,8 @@ async fn asset_and_storage_delta() -> anyhow::Result<()> {
 
     let tx_script_src = format!(
         r#"
-        use.mock::account
-        use.miden::output_note
+        use mock::account
+        use miden::protocol::output_note
 
         const MOCK_VALUE_SLOT0 = word("{mock_value_slot0}")
         const MOCK_MAP_SLOT = word("{mock_map_slot}")
@@ -848,9 +848,9 @@ async fn proven_tx_storage_maps_matches_executed_tx_for_new_account() -> anyhow:
 
     let code = format!(
         r#"
-      use.mock::account
+      use mock::account
 
-      const.MAP_SLOT=word("{map2_slot_name}")
+      const MAP_SLOT=word("{map2_slot_name}")
 
       begin
           # Update an existing key.
@@ -860,7 +860,7 @@ async fn proven_tx_storage_maps_matches_executed_tx_for_new_account() -> anyhow:
           # => [slot_id_prefix, slot_id_suffix, KEY, VALUE]
           call.account::set_map_item
 
-          exec.::std::sys::truncate_stack
+          exec.::miden::core::sys::truncate_stack
       end
       "#
     );
@@ -1104,12 +1104,12 @@ fn parse_tx_script(code: impl AsRef<str>) -> anyhow::Result<TransactionScript> {
 }
 
 const TEST_ACCOUNT_CONVENIENCE_WRAPPERS: &str = "
-      use.mock::account
-      use.miden::output_note
+      use mock::account
+      use miden::protocol::output_note
 
       #! Inputs:  [slot_id_prefix, slot_id_suffix, VALUE]
       #! Outputs: []
-      proc.set_item
+      proc set_item
           repeat.10 push.0 movdn.6 end
           # => [slot_id_prefix, slot_id_suffix, VALUE, pad(10)]
 
@@ -1121,7 +1121,7 @@ const TEST_ACCOUNT_CONVENIENCE_WRAPPERS: &str = "
 
       #! Inputs:  [slot_id_prefix, slot_id_suffix, KEY, VALUE]
       #! Outputs: []
-      proc.set_map_item
+      proc set_map_item
           repeat.6 push.0 movdn.10 end
           # => [index, KEY, VALUE, pad(6)]
 
@@ -1134,7 +1134,7 @@ const TEST_ACCOUNT_CONVENIENCE_WRAPPERS: &str = "
 
       #! Inputs:  [ASSET]
       #! Outputs: []
-      proc.create_note_with_asset
+      proc create_note_with_asset
           push.0.1.2.3           # recipient
           push.1                 # note_execution_hint
           push.2                 # note_type private
@@ -1154,7 +1154,7 @@ const TEST_ACCOUNT_CONVENIENCE_WRAPPERS: &str = "
 
       #! Inputs:  [tag, aux, note_type, execution_hint, RECIPIENT]
       #! Outputs: [note_idx]
-      proc.create_note
+      proc create_note
           repeat.8 push.0 movdn.8 end
           # => [tag, aux, note_type, execution_hint, RECIPIENT, pad(8)]
 
@@ -1167,7 +1167,7 @@ const TEST_ACCOUNT_CONVENIENCE_WRAPPERS: &str = "
 
       #! Inputs:  [ASSET, note_idx]
       #! Outputs: []
-      proc.move_asset_to_note
+      proc move_asset_to_note
           repeat.11 push.0 movdn.5 end
           # => [ASSET, note_idx, pad(11)]
 
@@ -1179,7 +1179,7 @@ const TEST_ACCOUNT_CONVENIENCE_WRAPPERS: &str = "
 
       #! Inputs:  [ASSET]
       #! Outputs: [ASSET']
-      proc.add_asset
+      proc add_asset
           repeat.12 push.0 movdn.4 end
           # => [ASSET, pad(12)]
 
@@ -1192,7 +1192,7 @@ const TEST_ACCOUNT_CONVENIENCE_WRAPPERS: &str = "
 
       #! Inputs:  [ASSET]
       #! Outputs: [ASSET]
-      proc.remove_asset
+      proc remove_asset
           repeat.12 push.0 movdn.4 end
           # => [ASSET, pad(12)]